Bring the repository in line with ess, torann and pyBlindOpt
The library was the odd one out: setup.cfg instead of PEP 621, Python >= 3.8,
no lint/type/dead-code gate, generated HTML committed to docs/, and a raster
logo where the others carry SVGs.

Packaging moves to pyproject.toml (setup.cfg deleted) with requires-python
>=3.12, project.urls, and the ruff/basedpyright/vulture configuration the other
three use. CI adopts ess's workflow verbatim - which says outright that it is
"the same gate as pyBlindOpt and torann" - matrixed over 3.12 and 3.14, plus
its docs and publish workflows and its pre-commit config. docs/ stops being
committed; the workflow is the only producer and docs_build/ is ignored.

Turning the gate on found real defects, not just style:

  evaluation.get_neighbourhood  previous_res was read in the else branch but
                                only assigned inside the loop, so t >= 1.0
                                raised NameError. Seeded before the loop.
  evaluation.compute_cost       the local `cost` shadowed the Metrics
                                parameter of the same name. Renamed to `value`.
  compute_global_segment_cost   removed. It called compute_cost(y, y_hat, cost)
                                against a (points, segment_errors, cost, cache)
                                signature - three arguments for four, and the
                                wrong ones - so it raised TypeError on every
                                call. No caller anywhere in src, test, demos or
                                examples; it cannot ever have worked.
  rdp.plot_frame                removed. Referenced `plt` from an import that
                                is commented out, so it raised NameError; its
                                only call site is itself commented out.
  rdp.min_point_rdp             t.sort() reordered the caller's list in place,
                                and with the mutable default that mutation
                                persisted between calls. Now sorted(t).
  lmethod.get_knee              annotated -> int while returning a 3-tuple.
  kneedle.knee/_knee            annotated -> int while returning None when no
                                peak is found.
  seven `callable` annotations  the builtin predicate, not a type. Callable.

The rest were annotation inaccuracies - numpy scalars declared as float, lists
declared as ndarray - corrected to match what the code returns.

Verified behaviour-neutral: the RDP reduction over all 20 trace x variant
combinations is byte-identical before and after, and the 133 tests stay green
throughout. ruff, basedpyright and vulture are all clean.

assets/logo.svg replaces media/knee.png, following pyBlindOpt's: 128x128, the
shared slate/blue/teal palette, prefers-color-scheme for dark, and title/desc/
aria-label. It draws the thing the library is for - a cost curve whose steep
descent and flat tail cross at the knee - and stays legible at favicon size,
which matters because docs.yml uses it as one.

README gets the header, badges and structure the other three carry, plus a
Determinism section for EPS_RANK. Every Python block in it was executed: they
run clean and print the values their comments claim.

pdoc builds 17 pages and the workflow's logo-resolution check passes on all of
them; python -m build and twine check both pass.

+++ src/kneeliverse/dfdt.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 
 '''
 The following module provides knee detection method
@@ -15,13 +14,14 @@ Copyright (c) 2021-2023 Stony Brook University
 Copyright (c) 2021-2023 The Research Foundation of SUNY
 '''
 
-import math
 import logging
+import math
+
 import numpy as np
 import uts.gradient as grad
 import uts.thresholding as thresh
-import kneeliverse.multi_knee as mk
 
+import kneeliverse.multi_knee as mk
 
 logger = logging.getLogger(__name__)
 
@@ -54,7 +54,7 @@ def get_knee_gradient(gradient: np.ndarray) -> int:
     t = thresh.isodata(gradient)
     diff = np.absolute(gradient - t)
     knee = np.argmin(diff[1:-1]) + 1
-    return knee
+    return int(knee)
 
 
 def knee(points: np.ndarray) -> int:
@@ -80,7 +80,7 @@ def knee(points: np.ndarray) -> int:
     while last_knee < knee and (len(x)-cutoff) > 2:
         last_knee = knee
         knee = get_knee_gradient(gradient[cutoff:]) + cutoff
-        cutoff = int(math.ceil(knee/2.0))
+        cutoff = math.ceil(knee/2.0)
 
     return knee
